% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juliaenvs.R
\name{activate_julia_env}
\alias{activate_julia_env}
\title{Set up R project to activate project-local Julia environments.}
\usage{
activate_julia_env()
}
\value{
TRUE invisibly, if function executes successfully. This funcion is
called entirely for its side effects.
}
\description{
Similar to the RStudio settings that allow the user to automatically
activate project-local Python environments, this function sets up the current
project to activate local Julia environments.
}
\examples{
\dontrun{
  # This function must be called within a directory that is an RStudio Project
  activate_julia_env()
}
}
