% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmolitorUtils-package.R
\docType{package}
\name{dmolitorUtils-package}
\alias{dmolitorUtils}
\alias{dmolitorUtils-package}
\title{dmolitorUtils: Personal Utility Functions}
\description{
A variety of utility functions that I use consistently.
}
\author{
\strong{Maintainer}: Daniel Molitor \email{molitdj97@gmail.com}

}
\keyword{internal}
