% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{holdout_metrics}
\alias{holdout_metrics}
\title{Calculate Model Metrics}
\usage{
holdout_metrics(truth, predicted, categorical = FALSE)
}
\arguments{
\item{truth}{A vector - the true outcome values.}

\item{predicted}{A vector - the predicted outcome values.}

\item{categorical}{A logical - Is the true outcome a categorical response
variable? This function only handles the binary case where \code{truth} is a
factor with levels 0 and 1.}
}
\value{
A tibble with calculated model metrics
}
\description{
Calculates a variety of standard model metrics, given the truth and
predicted values as vectors.
}
