% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{model_metrics}
\alias{model_metrics}
\title{Calculate Model Metrics}
\usage{
model_metrics(truth, predicted, categorical = FALSE)
}
\arguments{
\item{truth}{A vector - the true outcome values.}

\item{predicted}{A vector - the predicted outcome values.}

\item{categorical}{A logical - Is the true outcome a categorical response
variable? This function only handles the binary case where \code{truth} is a
factor with levels 0 and 1.}
}
\value{
A tibble with calculated model metrics
}
\description{
Calculates a variety of standard model metrics, given the truth and
predicted values as vectors.
}
\examples{
# Continuous outcome
mtcarslm <- lm(mpg ~ ., mtcars)
model_metrics(mtcars$mpg, mtcarslm$fitted.values)

# Binary outcome
iris$isSetosa <- factor(as.integer(iris$Species == "setosa"))
irisglm <- glm(isSetosa ~ Sepal.Length, iris, family = "binomial")
model_metrics(iris$isSetosa, irisglm$fitted.values, TRUE)

}
