% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{na_ref}
\alias{na_ref}
\title{Create Factor with NA as Reference}
\usage{
na_ref(var, ...)
}
\arguments{
\item{var}{Vector to convert to factor}

\item{...}{Additional arguments to pass to \code{\link[=factor]{factor()}}.}
}
\value{
\code{var} as a factor with NA as a reference level (if any exist).
}
\description{
Function that sets NA as factor reference level (if any NAs exist).
}
\examples{
no_na <- na_ref(1:10)
with_na <- na_ref(c(NA, 1:10))

levels(no_na)
levels(with_na)

}
