% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualenvs.R
\name{repl_python_virtualenv}
\alias{repl_python_virtualenv}
\title{Open Python REPL in virtual environment}
\usage{
repl_python_virtualenv(envdir)
}
\arguments{
\item{envdir}{The virtual environment directory.}
}
\value{
\code{NULL}. Called for its side effects.
}
\description{
\code{repl_python_virtualenv} Activates a Python virtual environment and
launches a Python REPL within that environment.
}
\examples{
\dontrun{
  repl_python_virtualenv("~/.venv/foo-bar/")
}

}
