% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{rmd_skeleton}
\alias{rmd_skeleton}
\title{RMarkdown Skeleton}
\usage{
rmd_skeleton(
  name = "Untitled",
  dir = here(),
  title = NULL,
  author = "Daniel Molitor",
  abstract = FALSE
)
}
\arguments{
\item{name}{A string - the name of the RMarkdown document.}

\item{dir}{The target directory.}

\item{title}{The title of the RMarkdown document. Optional argument.}

\item{author}{The author of the RMarkdown document. Optional argument.}

\item{abstract}{A logical value - whether to include an Abstract section.}
}
\value{
Returns \code{NULL} invisibly. It is called for its side effects
}
\description{
This function creates an RMarkdown skeleton with pretty defaults
}
\examples{
target_dir <- tempdir()
rmd_skeleton(name = "test",
             dir = target_dir,
             title = "Test Document")
file.exists(paste0(target_dir, "/test.Rmd"))
file.remove(paste0(target_dir, "/test.Rmd"))

}
