% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeway-formula.R
\name{threeway_formula}
\alias{threeway_formula}
\title{Convert a threeway formula into distinct sub-formulas}
\usage{
threeway_formula(form, drop.lhs = TRUE)
}
\arguments{
\item{form}{A \code{formula} or string that is coercible to one.}

\item{drop.lhs}{A logical indicating whether or not to drop the respective
LHS variables from their partner sub-formulas.}
}
\value{
A list of length 2, where the components are the sub-formulas of the
provided threeway formula.
}
\description{
A threeway formula is one with the form of \code{a ~ b ~ c} that implies the
following distinct sub-formulas: \code{a ~ b} and \code{a ~ c}. This function splits
such a threeway formula into its distinct sub-formulas.
}
\examples{
threeway_formula(z ~ y ~ w + w:x)
threeway_formula("z ~ y ~ w + w:x")
threeway_formula(mpg ~ carb ~ . - wt + gear:carb)
threeway_formula(mpg ~ carb ~ . - wt + gear:carb, drop.lhs = FALSE)

}
