% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualenvs.R
\name{win_dir}
\alias{win_dir}
\title{Translate filepaths to Windows style}
\usage{
win_dir(path, check.exists = FALSE)
}
\arguments{
\item{path}{Directory path.}

\item{check.exists}{A logical (not \code{NA}); check if directory exists.}
}
\value{
Directory path in Windows format.
}
\description{
Filepaths in the Windows OS typically use \verb{\\} as a separator. This function
strips out all forward slashes and replaces them with back slashes.
}
\examples{
target_dir <- gsub("\\\\\\\\", "/", tempdir())

# Convert directory path
win_dir(target_dir)
# Convert and check for directory existence
win_dir(target_dir, TRUE)

}
