% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmolitorUtils-package.R
\docType{package}
\name{dmolitorUtils-package}
\alias{dmolitorUtils}
\alias{dmolitorUtils-package}
\title{dmolitorUtils: Personal Utility Functions}
\description{
A variety of utility functions that I use consistently.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
