% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualenvs.R
\name{virtualenv_create_win}
\alias{virtualenv_create_win}
\title{Create Python virtual environment}
\usage{
virtualenv_create_win(envdir, envname, packages = NULL)
}
\arguments{
\item{envdir}{Directory to create virtual environment within.}

\item{envname}{Name of new virtual environment.}

\item{packages}{Additional python packages to install. When \code{NULL}, pip,
setuptools, and wheel are all installed by default.}
}
\value{
\code{NULL}. Called for side effects.
}
\description{
Create a new Python virtual environment and install updated versions of pip,
setuptools, and wheel.
}
